Use getPartsFromTemplate while loading urls
