+++ test/urls.js
@@ -4,17 +4,9 @@ module.exports = [
   {
     name: 'no-params',
     template: 'params/zero',
-    strings: ['params/zero'],
-    params: [],
   },
   {
     name: 'three-params',
     template: 'params/${first}-${second}/${third}/',
-    strings: ['params/', '-', '/', '/'],
-    params: [
-      'first',
-      'second',
-      'third',
-    ],
   },
 ];
