+++ test/utils.test.js
@@ -1,10 +1,4 @@
-function getPartsFromTemplate(template) {
-  const partRegex = /\$\{(\w+)\}/g;
-  const parts = template.split(partRegex);
-  const strings = parts.filter((element, index) => index % 2 === 0);
-  const params = parts.filter((element, index) => index % 2 === 1);
-  return { strings, params };
-}
+const { getPartsFromTemplate } = require('../src/utils');
 
 describe("getPartsFromTemplate", () => {
   it("should work without params", () => {
